Changed variable type of weight in User constructor from float to Double. Connection in connectToDB() now enforces Foreign Keys. Methods in UserDBOperations now open and close DB connection internally inserNewUser() now only takes a User as an argument Added updatedExistingUser() function Added commented out main method in UserDBOperations for manual testing

+++ team2fitness/src/main/java/seng202/group2/model/User.java
@@ -25,7 +25,7 @@ public class User {
 
 
 
-    public User(int id, String name, int age, double height, float weight){
+    public User(int id, String name, int age, double height, double weight){
         this.id = id;
         this.name = new SimpleStringProperty(name);
         this.age = new SimpleIntegerProperty(age);
